third_party/blink/web_tests/fast/css/local-file-name-requirements-expected.txt

CONSOLE WARNING: CSS cannot be loaded from `file:` URLs unless they end in a `.css` file extension.
This is a testharness.js-based test.
PASS `red.not-css` should not load
Harness: the test ran to completion.
